+++ packages/spacecat-shared-ims-client/src/auth.js
@@ -12,10 +12,10 @@
 
 import AWSXray from 'aws-xray-sdk';
 import { GetSecretValueCommand, SecretsManagerClient } from '@aws-sdk/client-secrets-manager';
-import { Site } from '@adobe/spacecat-shared-data-access';
-import { ImsPromiseClient } from '@adobe/spacecat-shared-ims-client';
-import { isString } from './functions.js';
-import { resolveCustomerSecretsName } from './helpers.js';
+import {
+  AUTHORING_TYPES, DELIVERY_TYPES, isString, resolveCustomerSecretsName,
+} from '@adobe/spacecat-shared-utils';
+import ImsPromiseClient from './clients/ims-promise-client.js';
 
 /**
  * @import {type Site} from "@adobe/spacecat-shared-data-access/src/models/site/index.js"
@@ -51,10 +51,10 @@ export async function getAccessToken(context, promiseToken) {
  * @throws {Error} - If secret is not found or token is missing
  */
 export async function retrievePageAuthentication(site, context, authOptions = {}) {
-  const CS_TYPES = [Site.AUTHORING_TYPES.CS, Site.AUTHORING_TYPES.CS_CW];
+  const CS_TYPES = [AUTHORING_TYPES.CS, AUTHORING_TYPES.CS_CW];
   if (site
     && (CS_TYPES.includes(site.getAuthoringType())
-      || site.getDeliveryType() === Site.DELIVERY_TYPES.AEM_CS)
+      || site.getDeliveryType() === DELIVERY_TYPES.AEM_CS)
     && authOptions.promiseToken) {
     return getAccessToken(context, authOptions.promiseToken.promise_token);
   }